refactor(auto-reply): move the foreground reply fence into its own module
check-lint flagged dispatch.ts (720) and dispatch.freshness.test.ts (1123)
over their max-lines caps after the bounded-wait change, and knip flagged
the test-only constant export. The fence owns enough behavior to be a
module: dispatch.ts drops to 482 effective lines, the wait-budget constant
becomes a real cross-module export, and the new bounded-wait tests live in
a colocated foreground-reply-fence.test.ts (dispatch.freshness.test.ts is
back to its pre-PR state).
